Optimized Network Handler by removing unneeded client-side code. Added Curseforge and Modrinth release tasks.

+++ common/src/main/java/de/markusbordihn/dailyrewards/menu/RewardCompactMenu.java
@@ -317,7 +317,7 @@ public class RewardCompactMenu extends RewardMenu {
                   specialSlotStartPositionX,
                   specialSlotStartPositionY));
         }
-        specialSlotStartPositionX += REWARD_SLOT_SIZE_X + REWARD_SLOT_SPACE_X;
+        specialSlotStartPositionX += (int) (REWARD_SLOT_SIZE_X + REWARD_SLOT_SPACE_X);
       }
     }
